refactor: update component imports and restructure paths for better organization
