+++ core/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import "../zap/lib/orpc/server";
+import "../zap/api/providers/orpc/server";
 import "./globals.css";
 
 import type { Metadata } from "next";
