+++ core/src/zap-old/components/common/footer.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 
-import { LatestBlogPosts } from "@/zap/components/blog/latest-blog-posts";
-import { ThemeSwitcher } from "@/zap/components/common/theme-switcher";
+import { LatestBlogPosts } from "@/zap/blog/components";
+import { ThemeSwitcher } from "@/zap-old/components/common/theme-switcher";
 
 const LEGAL_LINKS = [
   { href: "/terms-of-service", label: "Terms of Service" },
