+++ core/src/zap-old/components/landing/testimonials/marquee-row.tsx
@@ -1,6 +1,6 @@
 import { Marquee } from "@/components/magicui/marquee";
-import { ReviewCard } from "@/zap/components/landing/testimonials/review-card";
-import type { TESTIMONIALS } from "@/zap/data/landing";
+import { ReviewCard } from "@/zap-old/components/landing/testimonials/review-card";
+import { TESTIMONIALS } from "@/zap-old/data/landing";
 
 interface MarqueeRowProps {
   reviews: typeof TESTIMONIALS;
