+++ core/src/zap-old/hooks/users/use-number-of-users.ts
@@ -1,8 +1,8 @@
 "use client";
 import "client-only";
 
-import { useZapImmutable } from "@/zap/lib/api/hooks/use-zap-immutable";
-import { orpc } from "@/zap/lib/orpc/client";
+import { useZapImmutable } from "@/zap/api/hooks";
+import { orpc } from "@/zap/api/providers/orpc/client";
 
 export function useNumberOfUsers() {
   return useZapImmutable(orpc.users.getNumberOfUsers.queryOptions());
