+++ core/src/zap-old/rpc/procedures/example.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { authMiddleware, base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/api/handlers";
+import { withRpcHandler } from "@/zap/errors/handlers";
 
 export const example = base.use(authMiddleware).handler(
   withRpcHandler(({ context }) => {
