+++ core/src/zap/ai/components/index.ts
@@ -1,0 +1,1 @@
+export * from "@/zap/ai/components/settings-sheet";
