+++ core/src/zap/blog/components/index.ts
@@ -1,0 +1,1 @@
+export * from "@/zap/blog/components/latest-blog-posts";
